feat: `TimeExt`
