+++ core/src/ext/console.rs
@@ -1,23 +1,15 @@
-use std::time::Duration;
-
 use nova_vm::ecmascript::{
-    builtins::{
-        promise_objects::promise_abstract_operations::promise_capability_records::PromiseCapability,
-        ArgumentsList,
-    },
+    builtins::ArgumentsList,
     execution::{Agent, JsResult},
-    types::{GlobalValue, Value},
+    types::Value,
 };
 
-use crate::{
-    ext_interface::{Ext, ExtLoader},
-    HostData, MacroTask,
-};
+use crate::ext_interface::{Ext, ExtLoader};
 
 #[derive(Default)]
-pub struct ConsolExt;
+pub struct ConsoleExt;
 
-impl Ext for ConsolExt {
+impl Ext for ConsoleExt {
     fn load(&self, mut loader: ExtLoader) {
         loader.load_op("internal_read", Self::internal_read, 1);
         loader.load_op("internal_read_line", Self::internal_read_line, 1);
@@ -25,11 +17,10 @@ impl Ext for ConsolExt {
         loader.load_op("internal_write_line", Self::internal_write_line, 1);
         loader.load_op("debug", Self::debug, 1);
         loader.load_op("internal_exit", Self::internal_exit, 1);
-        loader.load_op("internal_sleep", Self::internal_sleep, 1); // TODO: Move to another ext
     }
 }
 
-impl ConsolExt {
+impl ConsoleExt {
     /// Debug function that prints the first argument to the console.
     fn debug(agent: &mut Agent, _this: Value, args: ArgumentsList) -> JsResult<Value> {
         if args.len() == 0 {
@@ -82,23 +73,5 @@ impl ConsolExt {
         }
         println!();
         Ok(Value::Undefined)
-    }
-
-    pub fn internal_sleep(agent: &mut Agent, _this: Value, args: ArgumentsList) -> JsResult<Value> {
-        let promise_capability = PromiseCapability::new(agent);
-        let time_ms = args[0].to_uint32(agent).unwrap();
-        let duration = Duration::from_millis(time_ms as u64);
-
-        let root_value = GlobalValue::new(agent, promise_capability.promise());
-        let host_data = agent.get_host_data();
-        let host_data: &HostData = host_data.downcast_ref().unwrap();
-        let macro_task_tx = host_data.macro_task_tx();
-
-        host_data.spawn_macro_task(async move {
-            tokio::time::sleep(duration).await;
-            macro_task_tx.send(MacroTask::ResolvePromise(root_value))
-        });
-
-        Ok(Value::Promise(promise_capability.promise()))
     }
 }
