+++ core/src/ext/mod.rs
@@ -1,5 +1,7 @@
 mod console;
 mod fs;
+mod time;
 
 pub use console::*;
 pub use fs::*;
+pub use time::*;
